Add files via upload
a proof of concept i will develop later tonight and aim to have a full system completed by midnight. I didnt want to buy a whole new NIC and take up space in my pc case. We all love a clean aesthetic. Anyway this it designed to plug into the power on pins on the motherboard. i will add more functionality later on. but anyway you SSH to the ESP, then input a command you want. This is meant to be for remote access needs so  i can turn on my desktop remotely and access its processing power that i dont have in the work office.